+++ examples/sentiment_analysis.rs
@@ -26,7 +26,7 @@ fn main() -> anyhow::Result<()> {
     ];
 
     //    Run model
-    let output = sentiment_classifier.predict(&input);
+    let output = sentiment_classifier.predict(input);
     for sentiment in output {
         println!("{:?}", sentiment);
     }